BE-66: change userId int to long

+++ src/main/java/az/gdg/msarticle/service/impl/CommentServiceImpl.java
@@ -50,7 +50,7 @@ public class CommentServiceImpl implements CommentService {
 
         String userId = (String) getAuthenticatedObject().getPrincipal();
         CommentEntity commentEntity = CommentMapper.INSTANCE.requestToEntity(commentRequest);
-        commentEntity.setUserId(Integer.parseInt(userId));
+        commentEntity.setUserId(Long.parseLong(userId));
 
         if (!commentRequest.getParentCommentId().isEmpty()) {
             commentEntity.setReply(true);
